+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Bitfinex parser: added new deposits/withdrawals export format.
 - Binance parser: added "Launchpool Airdrop", "HODLer Airdrops Distribution", "Megadrop Rewards", "Asset - Transfer", "Binance Card Spending", "Binance Card Cashback", "Crypto Box", "BNB Fee Deduction", "Transfer Between Main and Funding Wallet", "Transfer Between Main Account And Mining Account", "Transfer Between Main And Mining Account", "Fiat OCBS - Add Fiat and Fees", "Fiat Withdrawal", "Buy Crypto With Card", "Insurance Fund Refund" and "Cross Margin Liquidation - Repayment"  Operations.
 - Coinbase parser: added "Card Spend", "Credit", "Asset Migration", "Retail Eth2 Deprecation" transaction types.
+- Coinbase parser: added "Retail MGX DEX Buy" and "Retail MGX DEX Send" transaction types.
 ### Changed
 - Price tool: CoinDesk API deprecated.